Update the download.py script to also drop date-time formatting.
Currently the server returns datetime strings that cannot be parsed by
the rather strict go RFC3339 time.Time parser.

Why this solution?

- I already tried fixing the formatting before unmarshalling, but that
  code used regex and was rather brittle.
- The correct approach would be to use a custom wrapper type and
  implement the correct marshalling interface on it, but I'd rather not
  introduce yet more custom types and we'd still need an overlay
  x-go-type for the oapi-codegen part
- We could also create our own custom recursive unmarshalling alg that
  detects `time.Time` in the target type and applies custom parsing.
  This would be yet another thing to maintain and goes against general
  best pracices similar to point 1.

All of these solutions seem needless complicated for a problem that doesn't strictly need to be fixed right now.
The easiest, safest and most transparent solution is to simply use
strings, and let end users decide if they want to parse that string.
This is also most inline with our SDK approach for both go-denvr and
denvrpy.